Refactor code into smaller submodules
